Worked around iOS 6 Safari bug
Worked around a Safari bug in iOS 6 that prevents the screen from
redrawing properly for a few seconds upon initial page load. This
wasn't a bug in the tweening engine - it is an issue with Safari and
getting the WebkitTransform from the computed style of any element when
the page is initially loaded.