Major CSSPlugin update to allow 3D transforms, boxShadow, borderRadius, clip, and textShadow plus bug fixes and enhancements in misc. other files
- Adds ability for CSSPlugin to animate 3D transform properties like
rotationX, rotationY, rotationZ, z, perspective, transformPerspective,
shortRotationX, shortRotationY, shortRotationZ, etc.) as well as
borderRadius, boxShadow, textShadow, and clip (in browsers that support
those features)
- Adds ability for ScrollToPlugin to recognize a new "max" keyword to
scroll to the maximum position
- Improves the mechanics of the core "Ticker" and adds a workaround for
an iOS 6 bug that occasionally causes requestAnimationFrame updates not
to be called (the bug is in iOS 6 Safari, not GreenSock code)
- Adds ability for the "offset" parameter of the TimelineLite/Max
convenience methods like to(), from(), etc. to accept a label (string).
See docs for details - this can be a very convenient feature.
- Fixes bug in RoundPropsPlugin
- Fixes minor bugs in TimelineLite, TimelineMax, and the core Animation
class.
- Externalizes code necessary to integrate plugins like BezierPlugin
and ThrowPropsPlugin into CSSPlugin so that the kb impact of adding
more plugins in the future doesn't impact CSSPlugin
- Fixes overwrite behavior of className in CSSPlugin
- Prevents 2D transforms from overwriting gradient filter in old
versions of Internet Explorer