Fixed minor 3D transform issue
When transform:rotateY(180deg) was applied in css, it was interpreted
differently in various browsers. Like in IE10 it acted like rotationX,
rotationY, and rotationZ were all 180deg. Visually it's the same, but
for animation purposes it can be problematic. It had to do with
rounding issues in the browsers, so we worked around that.