Added EaseLookup, ability to recognize eases by name (String), fixed rare TimelineLite/Max bug and a CSSPlugin rotation issue
- Added EaseLookup to the EasePack so that you can find eases by name
(String). For example EaseLookup.find("Elastic.easeOut") or
EaseLookup.find("easeOutElastic")
- Added ability to define eases by name (String) like TweenLite.to(obj,
1, {x:10, ease:"Power4.easeOut"}). This only works for loaded eases, so
for example if you want "Elastic.easeOut" to work you need to load
EasePack (which is inside TweenMax already). "linear" and "swing" are
recognized too for maximum compatibility with jQuery. Define names in
the standard format like "Strong.easeOut" or reverse order and without
the dot (jQuery-style) "easeOutStrong"
- Fixed a minor issue that could (in very rare situations) cause an
element's starting rotation or skew to be interpreted between 180 and
-180 even if it exceeded those bounds from a previous tween.
- Now when a child is appended/inserted into a TimelineLite/Max, it
checks to see if it's already a child and if so, removes it before
calculating the position. Previously if you tried appending a child
that's already in the timeline, it could contaminate the duration.