Improved window.GreenSockGlobals and added @license
Now you can more easily accommodate loading (and using) multiple
versions of GSAP on the same page if necessary by defining a
GreenSockGlobals object on the window (completely optional). See the
comments toward the top of the uncompressed TweenLite.js file above the
"Definition" variable - it explains how to do it.