Minor update to jquery.gsap.js plugin, 0.1.2
In order to maximize compatibility, the plugin now falls back to the
native animate() function when a "step" function is defined. In this
case, none of the benefits of GSAP (speed or extra features) will be
available in the animation with the "step" function. We'd recommend
using an onUpdate instead if you can (or avoid both altogether).