2.0.0
When we released 1.20.5 on NPM with the new ES modules, it caused breaking changes in some build systems, so we reverted (1.20.6 on NPM) in less than 12 hours and then did a major version bump to avoid those problems. Sorry about that!! So although technically this is a major version bump, it’s virtually identical to 1.20.5. Please see the release notes for 1.20.5 at https://github.com/greensock/GreenSock-JS/releases/tag/1.20.5 for details about all the improvements.